Move a few lines to fix compilation warnings

+++ include/qrw/FootstepPlanner.hpp
@@ -42,7 +42,7 @@ public:
     ///
     ////////////////////////////////////////////////////////////////////////////////////////////////
     void initialize(double dt_in,
-                    double k_mpc_in,
+                    int k_mpc_in,
                     double T_mpc_in,
                     double h_ref_in,
                     Matrix34 const& shouldersIn,
